move lock cmd tests to own module
