+++ src/borg/testsuite/archiver/__init__.py
@@ -1935,10 +1935,6 @@ class ArchiverTestCase(ArchiverTestCaseBase):
         log = self.cmd(f"--repo={self.repository_location}", "--debug", "create", "test", "input")
         assert "security: read previous location" in log
 
-    def test_break_lock(self):
-        self.cmd(f"--repo={self.repository_location}", "rcreate", RK_ENCRYPTION)
-        self.cmd(f"--repo={self.repository_location}", "break-lock")
-
     def test_usage(self):
         self.cmd()
         self.cmd("-h")
@@ -1999,12 +1995,6 @@ class ArchiverTestCase(ArchiverTestCaseBase):
                 cache.commit()
         with pytest.raises(AssertionError):
             self.check_cache()
-
-    def test_with_lock(self):
-        self.cmd(f"--repo={self.repository_location}", "rcreate", RK_ENCRYPTION)
-        lock_path = os.path.join(self.repository_path, "lock.exclusive")
-        cmd = "python3", "-c", 'import os, sys; sys.exit(42 if os.path.exists("%s") else 23)' % lock_path
-        self.cmd(f"--repo={self.repository_location}", "with-lock", *cmd, fork=True, exit_code=42)
 
     def test_bad_filters(self):
         self.cmd(f"--repo={self.repository_location}", "rcreate", RK_ENCRYPTION)
